+++ src/plugins/javascript/moduleUtils.ts
@@ -104,22 +104,13 @@ export async function runScript(
   });
 
   const contextKeys = Object.keys(context);
-  const sourceAsync = `(async function userJS(exports, require, module, __filename, __dirname${
-    Object.keys(context).length > 0 ? `, ${Object.keys(context).join(', ')}` : ''
-  }){${io.fileProvider.EOL}${source}${io.fileProvider.EOL}})`;
+  const sourceAsync = `(async function userJS(exports, require, module, __filename, __dirname, context){ with(context){${io.fileProvider.EOL}${source}${io.fileProvider.EOL}}})`;
   const compiledWrapper = vm.runInThisContext(sourceAsync, {
     filename,
     lineOffset: options.lineOffset,
     displayErrors: true,
   });
-  await compiledWrapper.apply(context, [
-    mod.exports,
-    extendedRequire,
-    mod,
-    filename,
-    path.dirname(filename),
-    ...Object.values(context),
-  ]);
+  await compiledWrapper.apply(context, [mod.exports, extendedRequire, mod, filename, path.dirname(filename), context]);
 
   deleteVariables(contextKeys, context, options.deleteVariable);